renamed:    Info.java -> njusoftware/noComment/SamurAI/base/Info.java 	renamed:    Samurai.java -> njusoftware/noComment/SamurAI/base/Samurai.java
