+++ njusoftware/noComment/SamurAI/base/Samurai.java
@@ -2,8 +2,8 @@
 /**
  * @author clefz created at 2016/3/23
  *
- *         //请在此处标注编写者姓名 
- *         请在此处填写最终修改时间
+ *         //zqh st 
+ *         edited at 2016/3/27
  *         封装武士的状态
  */
 package njusoftware.noComment.SamurAI.base;
@@ -14,6 +14,44 @@ public class Samurai {
 	private int[] homePos = new int[2];
 	private int remainCurePeriod = 0;
 	private boolean isVisible = true;
+	
+	//封装武士的状态
+
+	public int[] getPos() {
+		return pos;
+	}
+
+	public void setPos(int[] pos) {
+		this.pos = pos;
+	}
+
+	public int[] getHomePos() {
+		return homePos;
+	}
+
+	public void setHomePos(int[] homePos) {
+		this.homePos = homePos;
+	}
+
+	public int getRemainCurePeriod() {
+		return remainCurePeriod;
+	}
+
+	public void setRemainCurePeriod(int remainCurePeriod) {
+		this.remainCurePeriod = remainCurePeriod;
+	}
+
+	public boolean isVisible() {
+		return isVisible;
+	}
+
+	public void setVisible(boolean isVisible) {
+		this.isVisible = isVisible;
+	}
+
+	public Weapons getWeapon() {
+		return weapon;
+	}
 
 	public Samurai(Weapons weapon) {
 		this.weapon = weapon;
